requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[argon2]
python-jose[cryptography]
email-validator
pydantic
requests
python-multipart
